# go/api_default.go
/*
 * Aliyun Api Gateway Swagger Sample
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.9
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

import (
	"encoding/json"
	"github.com/gorilla/mux"
	"log"
	"net/http"
)

func Case1(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	params := mux.Vars(r)
	id := params["userId"]
	log.Printf("user id %s", id)
	w.WriteHeader(http.StatusOK)
}

func Testpost(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(ResultOfGeneratePicturesVo{
		Model:     &GeneratePicturesVo{
			FailSize:    0,
			List:        []GeneratePictureVo{
				{
					Id:   0,
					Name: "test",
				},
			},
			SuccessSize: 0,
			TotalSize:   0,
		},
	})
}